% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data}
\alias{data}
\title{FTutils data}
\description{
Datasets related to the Danish Parlament
}
\details{
FTutils

Useful functions and tools for working with Danish parlamentary data.

Danish Parlamentary data
}
