% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folketingsvalg}
\alias{folketingsvalg}
\title{Datoer for folketingsvalg}
\format{
A character vector containing days for elections:
}
\usage{
folketingsvalg
}
\description{
Datoer for folketingsvalg
}
\keyword{datasets}
