% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partier}
\alias{partier}
\title{Partier}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{bogstav}{parti liste bogstav}
\item{navn}{partinavn}
...
}
}
\usage{
partier
}
\description{
Et datasæt der indeholder partinavne og listebogstaver
for partier i folketinget i perioden fra - til
}
\keyword{datasets}
