% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traktor}
\alias{traktor}
\title{Traktordata fra Maxs fædrende gaard}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}
}
\usage{
traktor
}
\description{
Traktordata fra Maxs fædrende gaard
}
\keyword{datasets}
