% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ugedage}
\alias{ugedage}
\title{Ugedagenes navne på dansk}
\format{
A character vector containing the danish weekdays:
}
\usage{
ugedage
}
\description{
Ugedagenes navne på dansk
}
\keyword{datasets}
