% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_assembly.R
\name{which_assembly}
\alias{which_assembly}
\title{which_assembly}
\arguments{
\item{x}{en teksstreng formatteret som YYYY-MM-DD}
}
\description{
Returnerer samling for en given dato
}
\details{
Returner samlinger

which_assembly
}
\examples{
# Returnerer 20201 for 1. samling i 2020:
which_assembly("2021-04-01")
}
