% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samlinger_dato}
\alias{samlinger_dato}
\title{"Datoer for folketingssamlinger"}
\format{
A tibble with 92 rows
\describe{
\item{Col1}{description of Col1}
\item{Col2}{description of Col2}
}
}
\source{
\url{https://www.gutenberg.org/files/524/524-0.txt}
}
\usage{
samlinger_dato
}
\description{
Et datasæt der indeholder start og slutdatoer for folketingssamlinger
mellem 6. oktober 1953 og 1. december 2021
}
\keyword{datasets}
