% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ugedage}
\alias{ugedage}
\title{Navne på ugedage.}
\format{
A data frame with five variables: \code{year}, \code{sex},
\code{name}, \code{n} and \code{prop} (\code{n} divided by total number
of applicants in that year, which means proportions are of people of
that sex with that name born in that year).
}
\usage{
ugedage
}
\description{
Full baby name data provided by the SSA. This includes all names with at
least 5 uses.
}
\keyword{datasets}
